Added hun.rlx -- right now it's almost the same as Mans's es-base.rle, I only modified the DELIMITERS clause.
Also, the following comments have accidentally been left out from the last commit:

hunmorph_to_apertium.foma works.

Fixed something in the rasskaz.hun.hunmorph file. The word I sent through
hunmorph had a hyphen (-) in it, and hunmorph does not handle this situation
well: instead of the regular xxx/YYY format, it returns xxx?YYY (and most likely
an incorrect analysis). For the time being, I changed the ? to / in the file,
but I am also going to look into how such words are supposed to be sent to
the program.

Added a few bugfixes to kr_to_apertium.lexc; it can convert all KR-codes
in the example file. Correctly? We'll have to see.